Fix `Gameplay_Key_Color.lua" script
The color should be set in OnAwake instead of OnStart because another script is trying to access the light color in OnStart as well. Resulting in Undefined Behavior which breaks the final door enigma.

Prior to https://github.com/Overload-Technologies/Overload/pull/437 there was no issues even though we didn't have any script ordering, it was just luck.